requirements.txt
python-telegram-bot==21.6
litellm==1.48.7
apscheduler==3.10.4
python-dotenv==1.0.1
